requirements.txt
websocket-client=0.57.0
requests~=2.24.0
grpcio~=1.48.0